Restored backward compatibility.
The only issue that was ignored was the one with changing the type of protected field for TaggedPdfReaderTool.java. This was done in the sake of automated porting. The similar issue with XmlDomWriter.java was fixed because whole this package is on manual porting. Also were added @deprecated annotations for fields and methods that are supposed to be removed at some point.

DEVSIX- 640

+++ layout/src/main/java/com/itextpdf/layout/element/Text.java
@@ -83,8 +83,8 @@ public class Text extends AbstractElement<Text> implements ILeafElement, IAccess
      * Gets the text rise.
      * @return the vertical distance from the text's default base line, as a float.
      */
-    public Float getTextRise() {
-        return this.<Float>getProperty(Property.TEXT_RISE);
+    public float getTextRise() {
+        return (float)this.<Float>getProperty(Property.TEXT_RISE);
     }
 
     /**
